@@ -21,6 +21,7 @@ private val wiremock = Testcontainer.wiremock
 @DisabledOnOs(OS.WINDOWS)
 class KtorAdapterTest {
 
+    @Suppress("UNCHECKED_CAST")
     class KtorBlockingFetcher(val ktorClient: HttpClient) : BlockingFetcher<HttpRequestBuilder> {
         override fun fetch(request: HttpRequestBuilder): Result = runBlocking {
             with(ktorClient.request(request)) {
@@ -30,7 +31,7 @@ class KtorAdapterTest {
                     contentType = contentType()?.toString(),
                     headers = headers.toMap().mapValues { it.value.firstOrNull().orEmpty() },
                     baseUri = request.url.toString(),
-                    cookies = listOf(setCookie()) as List<Cookie>
+                    cookies = listOf(setCookie()) as List<Cookie>,
                 )
             }
         }
@@ -39,13 +40,13 @@ class KtorAdapterTest {
             get() = HttpRequestBuilder()
     }
 
-    val KTOR_CLIENT = HttpClient(Apache)
+    private val ktorClient = HttpClient(Apache)
 
     @Test
     fun `dsl can skrape by url`() {
         wiremock.setupStub(path = "/example")
 
-        val result = skrape(KtorBlockingFetcher(KTOR_CLIENT)) {
+        val result = skrape(KtorBlockingFetcher(ktorClient)) {
             request {
                 url("${wiremock.httpUrl}/example")
             }
