@@ -12,7 +12,7 @@ public abstract class CssSelectable {
     internal abstract fun applySelector(rawCssSelector: String): List<DocElement>
 
     public fun <T> selection(cssSelector: String, init: CssSelector.() -> T): T =
-            CssSelector(rawCssSelector = cssSelector, doc = this).init()
+        CssSelector(rawCssSelector = cssSelector, doc = this).init()
 
     /**
      * Will create a CssSelector scope to calculate a css selector
@@ -20,12 +20,13 @@ public abstract class CssSelectable {
      * @return T
      */
     public operator fun <T> String.invoke(init: CssSelector.() -> T): T =
-            this@CssSelectable.selection(this, init)
+        this@CssSelectable.selection(this, init)
 
     public open fun makeDefault(cssSelector: String, relaxed: Boolean = false): DocElement {
         return DocElement(Element("${UUID.randomUUID()}"), relaxed)
     }
 
+    @Suppress("MaxLineLength")
     /**
      * Will pick all occurrences of elements that are matching the CSS-Selector
      * @see <a href="https://www.w3schools.com/cssref/css_selectors.asp">Overview of CSS-Selectors for further information.</a>
@@ -33,20 +34,21 @@ public abstract class CssSelectable {
      * @return T
      */
     public infix fun findAll(cssSelector: String): List<DocElement> =
-            this.applySelector(cssSelector)
+        this.applySelector(cssSelector)
 
     public fun findByIndex(index: Int, cssSelector: String = ""): DocElement =
-            findAll(cssSelector).getOrElse(index) { makeDefault(cssSelector) }
+        findAll(cssSelector).getOrElse(index) { makeDefault(cssSelector) }
 
     public operator fun Int.invoke(cssSelector: String = ""): DocElement =
-            findByIndex(this, cssSelector)
+        findByIndex(this, cssSelector)
 
     public fun findBySelectorMatching(regex: Regex): List<DocElement> =
-            this@CssSelectable.findAll("*").filter { it.ownCssSelector.matches(regex) }
+        this@CssSelectable.findAll("*").filter { it.ownCssSelector.matches(regex) }
 
     public operator fun Regex.invoke(): List<DocElement> =
-            findBySelectorMatching(this)
+        findBySelectorMatching(this)
 
+    @Suppress("MaxLineLength")
     /**
      * Will pick the first occurrence of an element that
      * is matching the CSS-Selector from a parsed document and invoke it to a lambda function.
@@ -55,20 +57,21 @@ public abstract class CssSelectable {
      * @return T
      */
     public infix fun findFirst(cssSelector: String): DocElement =
-            findByIndex(0, cssSelector)
+        findByIndex(0, cssSelector)
 
     public fun findSecond(cssSelector: String = ""): DocElement =
-            findByIndex(1, cssSelector)
+        findByIndex(1, cssSelector)
 
     public fun findThird(cssSelector: String = ""): DocElement =
-            findByIndex(2, cssSelector)
+        findByIndex(2, cssSelector)
 
     public fun findLast(cssSelector: String = ""): DocElement =
-            findAll(cssSelector).last()
+        findAll(cssSelector).last()
 
     public fun findSecondLast(cssSelector: String = ""): DocElement =
-            findAll(cssSelector).let { it.getOrElse(it.lastIndex -1) { makeDefault(cssSelector) } }
+        findAll(cssSelector).let { it.getOrElse(it.lastIndex - 1) { makeDefault(cssSelector) } }
 
+    @Suppress("MaxLineLength")
     /**
      * Will pick all occurrences of elements that are matching the CSS-Selector
      * @see <a href="https://www.w3schools.com/cssref/css_selectors.asp">Overview of CSS-Selectors for further information.</a>
@@ -76,20 +79,21 @@ public abstract class CssSelectable {
      * @return T
      */
     public fun <T> findAll(cssSelector: String = "", init: List<DocElement>.() -> T): T =
-            findAll(cssSelector).init()
+        findAll(cssSelector).init()
 
     public fun <T> findByIndex(index: Int, cssSelector: String = "", init: DocElement.() -> T): T =
-            findByIndex(index, cssSelector).init()
+        findByIndex(index, cssSelector).init()
 
     public operator fun <T> Int.invoke(cssSelector: String = "", init: DocElement.() -> T): T =
-            this(cssSelector).init()
+        this(cssSelector).init()
 
     public fun <T> findBySelectorMatching(regex: Regex, init: List<DocElement>.() -> T): T =
-            findBySelectorMatching(regex).init()
+        findBySelectorMatching(regex).init()
 
     public operator fun <T> Regex.invoke(init: List<DocElement>.() -> T): T =
-            this().init()
+        this().init()
 
+    @Suppress("MaxLineLength")
     /**
      * Will pick the first occurrence of an element that
      * is matching the CSS-Selector from a parsed document and invoke it to a lambda function.
@@ -98,17 +102,17 @@ public abstract class CssSelectable {
      * @return T
      */
     public fun <T> findFirst(cssSelector: String = "", init: DocElement.() -> T): T =
-            findFirst(cssSelector).init()
+        findFirst(cssSelector).init()
 
     public fun <T> findSecond(cssSelector: String = "", init: DocElement.() -> T): T =
-            findSecond(cssSelector).init()
+        findSecond(cssSelector).init()
 
     public fun <T> findThird(cssSelector: String = "", init: DocElement.() -> T): T =
-            findThird(cssSelector).init()
+        findThird(cssSelector).init()
 
     public fun <T> findLast(cssSelector: String = "", init: DocElement.() -> T): T =
-            findLast(cssSelector).init()
+        findLast(cssSelector).init()
 
     public fun <T> findSecondLast(cssSelector: String = "", init: DocElement.() -> T): T =
-            findSecondLast(cssSelector).init()
+        findSecondLast(cssSelector).init()
 }
