@@ -5,7 +5,7 @@ import it.skrape.selects.CssSelector
 
 /**
  * Will define a <del>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -18,7 +18,7 @@ public fun <T> CssSelectable.del(cssSelector: String = "", init: CssSelector.() 
 
 /**
  * Will define a <ins>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
