@@ -5,7 +5,7 @@ import it.skrape.selects.CssSelector
 
 /**
  * Will define a <applet>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -14,11 +14,11 @@ import it.skrape.selects.CssSelector
  * @return T
  */
 public fun <T> CssSelectable.applet(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("applet$cssSelector", init)
+    selection("applet$cssSelector", init)
 
 /**
  * Will define a <embed>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -27,11 +27,11 @@ public fun <T> CssSelectable.applet(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.embed(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("embed$cssSelector", init)
+    selection("embed$cssSelector", init)
 
 /**
  * Will define a <iframe>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -40,11 +40,11 @@ public fun <T> CssSelectable.embed(cssSelector: String = "", init: CssSelector.(
  * @return T
  */
 public fun <T> CssSelectable.iframe(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("iframe$cssSelector", init)
+    selection("iframe$cssSelector", init)
 
 /**
  * Will define a <noembed>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -53,11 +53,11 @@ public fun <T> CssSelectable.iframe(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.noembed(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("noembed$cssSelector", init)
+    selection("noembed$cssSelector", init)
 
 /**
  * Will define a <object>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -65,12 +65,13 @@ public fun <T> CssSelectable.noembed(cssSelector: String = "", init: CssSelector
  * @param cssSelector
  * @return T
  */
+@Suppress("FunctionNaming")
 public fun <T> CssSelectable.`object`(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("object$cssSelector", init)
+    selection("object$cssSelector", init)
 
 /**
  * Will define a <param>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -79,11 +80,11 @@ public fun <T> CssSelectable.`object`(cssSelector: String = "", init: CssSelecto
  * @return T
  */
 public fun <T> CssSelectable.param(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("param$cssSelector", init)
+    selection("param$cssSelector", init)
 
 /**
  * Will define a <picture>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -92,11 +93,11 @@ public fun <T> CssSelectable.param(cssSelector: String = "", init: CssSelector.(
  * @return T
  */
 public fun <T> CssSelectable.picture(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("picture$cssSelector", init)
+    selection("picture$cssSelector", init)
 
 /**
  * Will define a <source>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -105,4 +106,4 @@ public fun <T> CssSelectable.picture(cssSelector: String = "", init: CssSelector
  * @return T
  */
 public fun <T> CssSelectable.source(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("source$cssSelector", init)
+    selection("source$cssSelector", init)
