@@ -1,3 +1,5 @@
+@file:Suppress("TooManyFunctions")
+
 package it.skrape.selects.html5
 
 import it.skrape.selects.CssSelectable
@@ -5,7 +7,7 @@ import it.skrape.selects.CssSelector
 
 /**
  * Will define a <button>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -14,11 +16,11 @@ import it.skrape.selects.CssSelector
  * @return T
  */
 public fun <T> CssSelectable.button(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("button$cssSelector", init)
+    selection("button$cssSelector", init)
 
 /**
  * Will define a <datalist>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -27,11 +29,11 @@ public fun <T> CssSelectable.button(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.datalist(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("datalist$cssSelector", init)
+    selection("datalist$cssSelector", init)
 
 /**
  * Will define a <fieldset>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -40,11 +42,11 @@ public fun <T> CssSelectable.datalist(cssSelector: String = "", init: CssSelecto
  * @return T
  */
 public fun <T> CssSelectable.fieldset(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("fieldset$cssSelector", init)
+    selection("fieldset$cssSelector", init)
 
 /**
  * Will define a <form>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -53,11 +55,11 @@ public fun <T> CssSelectable.fieldset(cssSelector: String = "", init: CssSelecto
  * @return T
  */
 public fun <T> CssSelectable.form(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("form$cssSelector", init)
+    selection("form$cssSelector", init)
 
 /**
  * Will define a <input>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -66,11 +68,11 @@ public fun <T> CssSelectable.form(cssSelector: String = "", init: CssSelector.()
  * @return T
  */
 public fun <T> CssSelectable.input(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("input$cssSelector", init)
+    selection("input$cssSelector", init)
 
 /**
  * Will define a <label>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -79,11 +81,11 @@ public fun <T> CssSelectable.input(cssSelector: String = "", init: CssSelector.(
  * @return T
  */
 public fun <T> CssSelectable.label(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("label$cssSelector", init)
+    selection("label$cssSelector", init)
 
 /**
  * Will define a <legend>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -92,11 +94,11 @@ public fun <T> CssSelectable.label(cssSelector: String = "", init: CssSelector.(
  * @return T
  */
 public fun <T> CssSelectable.legend(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("legend$cssSelector", init)
+    selection("legend$cssSelector", init)
 
 /**
  * Will define a <meter>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -105,11 +107,11 @@ public fun <T> CssSelectable.legend(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.meter(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("meter$cssSelector", init)
+    selection("meter$cssSelector", init)
 
 /**
  * Will define a <optgroup>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -118,11 +120,11 @@ public fun <T> CssSelectable.meter(cssSelector: String = "", init: CssSelector.(
  * @return T
  */
 public fun <T> CssSelectable.optgroup(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("optgroup$cssSelector", init)
+    selection("optgroup$cssSelector", init)
 
 /**
  * Will define a <option>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -131,11 +133,11 @@ public fun <T> CssSelectable.optgroup(cssSelector: String = "", init: CssSelecto
  * @return T
  */
 public fun <T> CssSelectable.option(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("option$cssSelector", init)
+    selection("option$cssSelector", init)
 
 /**
  * Will define a <output>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -144,11 +146,11 @@ public fun <T> CssSelectable.option(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.output(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("output$cssSelector", init)
+    selection("output$cssSelector", init)
 
 /**
  * Will define a <progress>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -157,11 +159,11 @@ public fun <T> CssSelectable.output(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.progress(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("progress$cssSelector", init)
+    selection("progress$cssSelector", init)
 
 /**
  * Will define a <select>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -170,11 +172,11 @@ public fun <T> CssSelectable.progress(cssSelector: String = "", init: CssSelecto
  * @return T
  */
 public fun <T> CssSelectable.select(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("select$cssSelector", init)
+    selection("select$cssSelector", init)
 
 /**
  * Will define a <textarea>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -183,4 +185,4 @@ public fun <T> CssSelectable.select(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.textarea(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("textarea$cssSelector", init)
+    selection("textarea$cssSelector", init)
