@@ -5,7 +5,7 @@ import it.skrape.selects.CssSelector
 
 /**
  * Will define a <area>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -18,7 +18,7 @@ public fun <T> CssSelectable.area(cssSelector: String = "", init: CssSelector.()
 
 /**
  * Will define a <audio>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -31,7 +31,7 @@ public fun <T> CssSelectable.audio(cssSelector: String = "", init: CssSelector.(
 
 /**
  * Will define a <img>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -44,7 +44,7 @@ public fun <T> CssSelectable.img(cssSelector: String = "", init: CssSelector.() 
 
 /**
  * Will define a <map>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -57,7 +57,7 @@ public fun <T> CssSelectable.map(cssSelector: String = "", init: CssSelector.() 
 
 /**
  * Will define a <track>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -70,7 +70,7 @@ public fun <T> CssSelectable.track(cssSelector: String = "", init: CssSelector.(
 
 /**
  * Will define a <video>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
