@@ -1,3 +1,5 @@
+@file:Suppress("TooManyFunctions")
+
 package it.skrape.selects.html5
 
 import it.skrape.selects.CssSelectable
@@ -5,7 +7,7 @@ import it.skrape.selects.CssSelector
 
 /**
  * Will define a <blockquote>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -14,11 +16,11 @@ import it.skrape.selects.CssSelector
  * @return T
  */
 public fun <T> CssSelectable.blockquote(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("blockquote$cssSelector", init)
+    selection("blockquote$cssSelector", init)
 
 /**
  * Will define a <dd>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -27,11 +29,11 @@ public fun <T> CssSelectable.blockquote(cssSelector: String = "", init: CssSelec
  * @return T
  */
 public fun <T> CssSelectable.dd(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("dd$cssSelector", init)
+    selection("dd$cssSelector", init)
 
 /**
  * Will define a <dir>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -40,11 +42,11 @@ public fun <T> CssSelectable.dd(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.dir(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("dir$cssSelector", init)
+    selection("dir$cssSelector", init)
 
 /**
  * Will define a <dl>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -53,11 +55,11 @@ public fun <T> CssSelectable.dir(cssSelector: String = "", init: CssSelector.() 
  * @return T
  */
 public fun <T> CssSelectable.dl(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("dl$cssSelector", init)
+    selection("dl$cssSelector", init)
 
 /**
  * Will define a <dt>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -66,11 +68,11 @@ public fun <T> CssSelectable.dl(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.dt(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("dt$cssSelector", init)
+    selection("dt$cssSelector", init)
 
 /**
  * Will define a <figcaption>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -79,11 +81,11 @@ public fun <T> CssSelectable.dt(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.figcaption(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("figcaption$cssSelector", init)
+    selection("figcaption$cssSelector", init)
 
 /**
  * Will define a <figure>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -92,11 +94,11 @@ public fun <T> CssSelectable.figcaption(cssSelector: String = "", init: CssSelec
  * @return T
  */
 public fun <T> CssSelectable.figure(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("figure$cssSelector", init)
+    selection("figure$cssSelector", init)
 
 /**
  * Will define a <hr>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -105,11 +107,11 @@ public fun <T> CssSelectable.figure(cssSelector: String = "", init: CssSelector.
  * @return T
  */
 public fun <T> CssSelectable.hr(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("hr$cssSelector", init)
+    selection("hr$cssSelector", init)
 
 /**
  * Will define a <li>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -118,11 +120,11 @@ public fun <T> CssSelectable.hr(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.li(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("li$cssSelector", init)
+    selection("li$cssSelector", init)
 
 /**
  * Will define a <ol>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -131,11 +133,11 @@ public fun <T> CssSelectable.li(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.ol(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("ol$cssSelector", init)
+    selection("ol$cssSelector", init)
 
 /**
  * Will define a <ul>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -144,11 +146,11 @@ public fun <T> CssSelectable.ol(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.ul(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("ul$cssSelector", init)
+    selection("ul$cssSelector", init)
 
 /**
  * Will define a <p>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -157,11 +159,11 @@ public fun <T> CssSelectable.ul(cssSelector: String = "", init: CssSelector.() -
  * @return T
  */
 public fun <T> CssSelectable.p(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("p$cssSelector", init)
+    selection("p$cssSelector", init)
 
 /**
  * Will define a <pre>-tags css query selector.
- * By default it will just be the specific tag-name.
+ * By default, it will just be the specific tag-name.
  * It is possible to define a more concrete selector by using the provided fields of the CssSelector object or
  * by passing a raw css query selector as parameter.
  * If a selector is passed as parameter as well as be defined via CssSelector fields, they will be merged.
@@ -170,4 +172,4 @@ public fun <T> CssSelectable.p(cssSelector: String = "", init: CssSelector.() ->
  * @return T
  */
 public fun <T> CssSelectable.pre(cssSelector: String = "", init: CssSelector.() -> T): T =
-        selection("pre$cssSelector", init)
+    selection("pre$cssSelector", init)
