@@ -9,7 +9,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate selector from raw css selector param`() {
         val cssSelector = CssSelector(
-                rawCssSelector = "#foo.bar"
+            rawCssSelector = "#foo.bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foo.bar")
     }
@@ -17,7 +17,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate selector from raw css selector param with spaces`() {
         val cssSelector = CssSelector(
-                rawCssSelector = "div #foo.bar"
+            rawCssSelector = "div #foo.bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("div #foo.bar")
     }
@@ -25,7 +25,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate class selector from element`() {
         val cssSelector = CssSelector(
-                withClass = "foo"
+            withClass = "foo",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo(".foo")
     }
@@ -33,7 +33,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate list of classes selector from element`() {
         val cssSelector = CssSelector(
-                withClass = "foo" and "bar"
+            withClass = "foo" and "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo(".foo.bar")
     }
@@ -41,7 +41,7 @@ class CssSelectorTest {
     @Test
     fun `whitespaces in class names will be kicked`() {
         val cssSelector = CssSelector(
-            withClass = "   foobar " and " foo " and " bar "
+            withClass = "   foobar " and " foo " and " bar ",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo(".foobar.foo.bar")
     }
@@ -49,19 +49,20 @@ class CssSelectorTest {
     @Test
     fun `can calculate selector and be relaxed on miss-leading spaces`() {
         val cssSelector = CssSelector(
-                rawCssSelector = "   div.foo span   ",
-                withClass = "   foobar " and " foo " and " bar ",
-                withAttribute = "   foooo " to "  bar   ",
-                withAttributes = "fizz" to "buzz" and Pair("skrape", "it"),
-                withAttributeKeys = listOf("key1", "key2")
+            rawCssSelector = "   div.foo span   ",
+            withClass = "   foobar " and " foo " and " bar ",
+            withAttribute = "   foooo " to "  bar   ",
+            withAttributes = "fizz" to "buzz" and Pair("skrape", "it"),
+            withAttributeKeys = listOf("key1", "key2"),
         ).toCssSelector
-        expectThat(cssSelector).isEqualTo("div.foo span.foobar.foo.bar['key1']['key2'][foooo='  bar   '][fizz='buzz'][skrape='it']")
+        expectThat(cssSelector)
+            .isEqualTo("div.foo span.foobar.foo.bar['key1']['key2'][foooo='  bar   '][fizz='buzz'][skrape='it']")
     }
 
     @Test
     fun `can calculate id selector from element`() {
         val cssSelector = CssSelector(
-                withId = "foo"
+            withId = "foo",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foo")
     }
@@ -69,7 +70,7 @@ class CssSelectorTest {
     @Test
     fun `whitespaces in id will be kicked`() {
         val cssSelector = CssSelector(
-            withId = " f o o  "
+            withId = " f o o  ",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foo")
     }
@@ -77,7 +78,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate attributeKey selector from element`() {
         val cssSelector = CssSelector(
-            withAttributeKey = "foo"
+            withAttributeKey = "foo",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("[foo]")
     }
@@ -85,7 +86,7 @@ class CssSelectorTest {
     @Test
     fun `whitespaces in attributeKey will be kicked`() {
         val cssSelector = CssSelector(
-            withAttributeKey = " f o o  "
+            withAttributeKey = " f o o  ",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("[foo]")
     }
@@ -93,7 +94,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate attribute selector from element`() {
         val cssSelector = CssSelector(
-                withAttribute = "foo" to "bar"
+            withAttribute = "foo" to "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("[foo='bar']")
     }
@@ -101,7 +102,7 @@ class CssSelectorTest {
     @Test
     fun `can calculate attribute selector with attribute value containing whitespaces`() {
         val cssSelector = CssSelector(
-            withAttribute = "foo" to " bar foobar  "
+            withAttribute = "foo" to " bar foobar  ",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("[foo=' bar foobar  ']")
     }
@@ -109,7 +110,7 @@ class CssSelectorTest {
     @Test
     fun `whitespaces in attribute key will be kicked`() {
         val cssSelector = CssSelector(
-            withAttribute = "   f o o  " to "bar"
+            withAttribute = "   f o o  " to "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("[foo='bar']")
     }
@@ -117,8 +118,8 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of id and class selector from element`() {
         val cssSelector = CssSelector(
-                withClass = "bar",
-                withId = "foo"
+            withClass = "bar",
+            withId = "foo",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foo.bar")
     }
@@ -126,8 +127,8 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of id and attributeKey selector from element`() {
         val cssSelector = CssSelector(
-                withAttributeKey = "bar",
-                withId = "foo"
+            withAttributeKey = "bar",
+            withId = "foo",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foo[bar]")
     }
@@ -135,8 +136,8 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of id and attribute selector from element`() {
         val cssSelector = CssSelector(
-                withAttribute = "foo" to "bar",
-                withId = "foobar"
+            withAttribute = "foo" to "bar",
+            withId = "foobar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foobar[foo='bar']")
     }
@@ -144,8 +145,8 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of attribute and class selector from element`() {
         val cssSelector = CssSelector(
-                withClass = "foobar",
-                withAttribute = "foo" to "bar"
+            withClass = "foobar",
+            withAttribute = "foo" to "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo(".foobar[foo='bar']")
     }
@@ -153,8 +154,8 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of attribute and attributeKey selector from element`() {
         val cssSelector = CssSelector(
-                withAttributeKey = "foobar",
-                withAttribute = "foo" to "bar"
+            withAttributeKey = "foobar",
+            withAttribute = "foo" to "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("[foobar][foo='bar']")
     }
@@ -162,9 +163,9 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of id, attributeKey and class selector from element`() {
         val cssSelector = CssSelector(
-                withId = "foobar",
-                withAttributeKey = "foo",
-                withClass = "bar"
+            withId = "foobar",
+            withAttributeKey = "foo",
+            withClass = "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#foobar.bar[foo]")
     }
@@ -172,9 +173,9 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of id, attribute and attributeKey selector from element`() {
         val cssSelector = CssSelector(
-                withId = "fb",
-                withAttributeKey = "foobar",
-                withAttribute = "foo" to "bar"
+            withId = "fb",
+            withAttributeKey = "foobar",
+            withAttribute = "foo" to "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#fb[foobar][foo='bar']")
     }
@@ -182,9 +183,9 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of id, attribute and class selector from element`() {
         val cssSelector = CssSelector(
-                withId = "fb",
-                withClass = "foobar",
-                withAttribute = "foo" to "bar"
+            withId = "fb",
+            withClass = "foobar",
+            withAttribute = "foo" to "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("#fb.foobar[foo='bar']")
     }
@@ -192,8 +193,8 @@ class CssSelectorTest {
     @Test
     fun `can calculate combination of attributeKey and class selector from element`() {
         val cssSelector = CssSelector(
-                withAttributeKey = "foo",
-                withClass = "bar"
+            withAttributeKey = "foo",
+            withClass = "bar",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo(".bar[foo]")
     }
@@ -201,11 +202,11 @@ class CssSelectorTest {
     @Test
     fun `can calculate complex css selector from element`() {
         val cssSelector = CssSelector(
-                rawCssSelector = "div span a",
-                withClass = "bar",
-                withId = "foo",
-                withAttributeKey = "foobar",
-                withAttribute = "fizz" to "buzz"
+            rawCssSelector = "div span a",
+            withClass = "bar",
+            withId = "foo",
+            withAttributeKey = "foobar",
+            withAttribute = "fizz" to "buzz",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("div span a#foo.bar[foobar][fizz='buzz']")
     }
@@ -213,21 +214,23 @@ class CssSelectorTest {
     @Test
     fun `parameter selector will be merged with element selector`() {
         val cssSelector = CssSelector(
-                withClass = "bar",
-                rawCssSelector = "foo"
+            withClass = "bar",
+            rawCssSelector = "foo",
         ).toCssSelector
         expectThat(cssSelector).isEqualTo("foo.bar")
     }
 
     @Test
     fun `'toString()' implementation returns calculated css selector`() {
-        val cssSelector = "${CssSelector(
-            rawCssSelector = "div span a",
-            withClass = "bar",
-            withId = "foo",
-            withAttributeKey = "foobar",
-            withAttribute = "fizz" to "buzz"
-        )}"
+        val cssSelector = "${
+            CssSelector(
+                rawCssSelector = "div span a",
+                withClass = "bar",
+                withId = "foo",
+                withAttributeKey = "foobar",
+                withAttribute = "fizz" to "buzz",
+            )
+        }"
         expectThat(cssSelector).isEqualTo("div span a#foo.bar[foobar][fizz='buzz']")
     }
 }
