@@ -1,3 +1,5 @@
+@file:Suppress("LargeClass")
+
 package it.skrape.selects
 
 import aValidDocument
@@ -61,7 +63,7 @@ class DocElementKtTest {
         val someElements = listOf(aValidElement, aValidElement)
         expectThat(someElements.eachText).containsExactly(
             "divs text headline paragraph foo bar fizz buzz",
-            "divs text headline paragraph foo bar fizz buzz"
+            "divs text headline paragraph foo bar fizz buzz",
         )
     }
 
@@ -77,7 +79,7 @@ class DocElementKtTest {
         paragraph
             foo bar
             fizz buzz
-        """
+        """,
         )
     }
 
@@ -88,7 +90,7 @@ class DocElementKtTest {
             divs text 
             <h2 class="welcome" disabled>headline</h2> 
             <p class="fancy">paragraph <span>foo <b>bar</b></span> <span>fizz <b id="xxx">buzz</b></span> </p>
-            """.trimIndent()
+            """.trimIndent(),
         )
     }
 
@@ -101,7 +103,7 @@ class DocElementKtTest {
              <h2 class="welcome" disabled>headline</h2> 
              <p class="fancy">paragraph <span>foo <b>bar</b></span> <span>fizz <b id="xxx">buzz</b></span> </p>
             </div>
-            """.trimIndent()
+            """.trimIndent(),
         )
     }
 
@@ -173,7 +175,7 @@ class DocElementKtTest {
     }
 
     @Test
-    fun `can find all elements within this element by selector and invoke them to lambda that will return generic value`() {
+    fun `can find all elements within this element by selector and invoke them to lambda will return generic value`() {
         val text = aValidElement.findAll(".welcome") {
             expectThat(this).hasSize(1)
             text
@@ -182,7 +184,7 @@ class DocElementKtTest {
     }
 
     @Test
-    fun `can find first element within this element by selector and invoke them to lambda that will return generic value`() {
+    fun `can find first element within this element by selector and invoke them to lambda will return generic value`() {
         val text = aValidElement.findFirst(".welcome") {
             text
         }
@@ -213,12 +215,14 @@ class DocElementKtTest {
 
     @Test
     fun `can get ownCssSelector`() {
-        val elementWithId = DocElement(anElement.apply {
-            attr("id", "bazinga")
-            attr("key-only-attr", "")
-        })
+        val elementWithId = DocElement(
+            anElement.apply {
+                attr("id", "bazinga")
+                attr("key-only-attr", "")
+            },
+        )
         expectThat(elementWithId.ownCssSelector).isEqualTo(
-            "div#bazinga.clazz.klass['key-only-attr'][foo='bar'][fizz='buzz'][data-foo='foobar']"
+            "div#bazinga.clazz.klass['key-only-attr'][foo='bar'][fizz='buzz'][data-foo='foobar']",
         )
     }
 
@@ -266,22 +270,28 @@ class DocElementKtTest {
                 "class" to "clazz klass",
                 "foo" to "bar",
                 "fizz" to "buzz",
-                "data-foo" to "foobar"
-            )
+                "data-foo" to "foobar",
+            ),
         )
     }
 
     @Test
     fun `can get all attribute keys of an element`() {
         expectThat(aValidElement.attributeKeys).containsExactly(
-            "class", "foo", "fizz", "data-foo"
+            "class",
+            "foo",
+            "fizz",
+            "data-foo",
         )
     }
 
     @Test
     fun `can get all attribute values of an element`() {
         expectThat(aValidElement.attributeValues).containsExactly(
-            "clazz klass", "bar", "buzz", "foobar"
+            "clazz klass",
+            "bar",
+            "buzz",
+            "foobar",
         )
     }
 
@@ -293,15 +303,15 @@ class DocElementKtTest {
                 "class" to "clazz klass",
                 "foo" to "bar",
                 "fizz" to "buzz",
-                "data-foo" to "foobar"
-            )
+                "data-foo" to "foobar",
+            ),
         )
     }
 
     @Test
     fun `can get all data attributes of an elements`() {
         expectThat(aValidElement.dataAttributes).isEqualTo(
-            mapOf("data-foo" to "foobar")
+            mapOf("data-foo" to "foobar"),
         )
     }
 
@@ -309,7 +319,7 @@ class DocElementKtTest {
     fun `can get all data attributes from a list of elements`() {
         val someElements = listOf(aValidElement, aValidElement)
         expectThat(someElements.eachDataAttribute).isEqualTo(
-            mapOf("data-foo" to "foobar")
+            mapOf("data-foo" to "foobar"),
         )
     }
 
@@ -338,7 +348,6 @@ class DocElementKtTest {
             <div class="my-class"><h1 class="welcome">first headline</h1></div>
             <div class="my-class"><h1 class="welcome">second headline</h1></div>
         """.trimIndent()
-
 
         val text = htmlDocument(markup) {
             div {
@@ -425,7 +434,7 @@ class DocElementKtTest {
              <h2 class="welcome" disabled>headline</h2> 
              <p class="fancy">paragraph <span>foo <b>bar</b></span> <span>fizz <b id="xxx">buzz</b></span> </p>
             </div>
-            """.trimIndent()
+            """.trimIndent(),
         )
     }
 
@@ -449,7 +458,6 @@ class DocElementKtTest {
 
     @Test
     fun `can conveniently get all links as map of text and href from list of DocElement`() {
-
         val links = aValidDocument(aValidMarkupWithLinks) {
             a {
                 findAll {
@@ -464,15 +472,14 @@ class DocElementKtTest {
                 "relative link" to "/relative",
                 "modal" to "#modal",
                 "nested link" to "http://nested.link",
-                "i'm an anchor" to "http://some.link"
-            )
+                "i'm an anchor" to "http://some.link",
+            ),
         )
     }
 
     @Test
     @Disabled("TODO: make eachLink also filter child notes")
     fun `can conveniently get all links as map of text and href from list of DocElement and its children`() {
-
         val links = aValidDocument(aValidMarkupWithLinks) {
             div {
                 findAll {
@@ -488,10 +495,9 @@ class DocElementKtTest {
                 "fizzbuzz" to "http://fizz.buzz",
                 "schnitzel" to "http://schnitzel.de",
                 "nested link" to "http://nested.link",
-                "i'm an anchor" to "http://some.link"
-            )
+                "i'm an anchor" to "http://some.link",
+            ),
         )
-
     }
 
     @Test
@@ -505,7 +511,6 @@ class DocElementKtTest {
         }
 
         expectThat(links).isEqualTo(mapOf("foobar" to "http://foo.bar"))
-
     }
 
     @Test
@@ -521,10 +526,9 @@ class DocElementKtTest {
                 "relative link" to "/relative",
                 "modal" to "#modal",
                 "nested link" to "http://nested.link",
-                "i'm an anchor" to "http://some.link"
-            )
+                "i'm an anchor" to "http://some.link",
+            ),
         )
-
     }
 
     @Test
@@ -544,8 +548,8 @@ class DocElementKtTest {
                 "foobar" to "http://foo.bar",
                 "" to "http://fizz.buzz",
                 "yummi" to "http://schnitzel.de",
-                "nested" to "http://nested.image"
-            )
+                "nested" to "http://nested.image",
+            ),
         )
     }
 
@@ -558,7 +562,7 @@ class DocElementKtTest {
                 "/relative",
                 "#modal",
                 "http://nested.link",
-                "http://some.link"
+                "http://some.link",
             )
             a {
                 findAll {
@@ -567,7 +571,7 @@ class DocElementKtTest {
                         "/relative",
                         "#modal",
                         "http://nested.link",
-                        "http://some.link"
+                        "http://some.link",
                     )
                 }
             }
@@ -582,7 +586,7 @@ class DocElementKtTest {
                 "http://foo.bar",
                 "http://fizz.buzz",
                 "http://schnitzel.de",
-                "http://nested.image"
+                "http://nested.image",
             )
             img {
                 findAll {
@@ -590,7 +594,7 @@ class DocElementKtTest {
                         "http://foo.bar",
                         "http://fizz.buzz",
                         "http://schnitzel.de",
-                        "http://nested.image"
+                        "http://nested.image",
                     )
                 }
             }
@@ -601,12 +605,12 @@ class DocElementKtTest {
     fun `can conveniently get all custom attributes`() {
         aValidDocument(aValidMarkupWithPictures) {
             expectThat(eachAttribute("rel")).containsExactly(
-                "shortcut icon"
+                "shortcut icon",
             )
             link {
                 findAll {
                     expectThat(eachAttribute("rel")).containsExactly(
-                        "shortcut icon"
+                        "shortcut icon",
                     )
                 }
             }
@@ -620,7 +624,6 @@ class DocElementKtTest {
 
     @Test
     fun `can conveniently get all images as map of alt-text and src from list of DocElement`() {
-
         val pictures = aValidDocument(aValidMarkupWithPictures) {
             img {
                 findAll {
@@ -634,15 +637,14 @@ class DocElementKtTest {
                 "foobar" to "http://foo.bar",
                 "" to "http://fizz.buzz",
                 "yummi" to "http://schnitzel.de",
-                "nested" to "http://nested.image"
-            )
+                "nested" to "http://nested.image",
+            ),
         )
     }
 
     @Test
     @Disabled("TODO: make eachImage also filter child notes")
     fun `can conveniently get all images as map of alt-text and src from list of DocElement and its children`() {
-
         val pictures = aValidDocument(aValidMarkupWithPictures) {
             div {
                 findAll {
@@ -656,10 +658,9 @@ class DocElementKtTest {
                 "foobar" to "http://foo.bar",
                 "" to "http://fizz.buzz",
                 "yummi" to "http://schnitzel.de",
-                "nested" to "http://nested.image"
-            )
+                "nested" to "http://nested.image",
+            ),
         )
-
     }
 
     @Test
@@ -673,7 +674,6 @@ class DocElementKtTest {
         }
 
         expectThat(pictures).isEqualTo(mapOf("foobar" to "http://foo.bar"))
-
     }
 
     @Test
@@ -687,10 +687,9 @@ class DocElementKtTest {
                 "foobar" to "http://foo.bar",
                 "" to "http://fizz.buzz",
                 "yummi" to "http://schnitzel.de",
-                "nested" to "http://nested.image"
-            )
+                "nested" to "http://nested.image",
+            ),
         )
-
     }
 
     @Test
@@ -698,13 +697,12 @@ class DocElementKtTest {
         expectThat(aValidElement.element).isA<Element>()
     }
 
-
     @Test
     fun `children, parents, siblings and allElements of relaxed element are also relaxed`() {
         htmlDocument("<div><a>skrape.it</a></div><div>sibling</div>") {
             findFirst("div") {
                 expectThat(
-                    children[0].relaxed || parents[0].relaxed || siblings[0].relaxed || allElements[0].relaxed
+                    children[0].relaxed || parents[0].relaxed || siblings[0].relaxed || allElements[0].relaxed,
                 ).isEqualTo(false)
             }
 
@@ -712,7 +710,7 @@ class DocElementKtTest {
 
             findFirst("div") {
                 expectThat(
-                    children[0].relaxed && parents[0].relaxed && siblings[0].relaxed && allElements[0].relaxed
+                    children[0].relaxed && parents[0].relaxed && siblings[0].relaxed && allElements[0].relaxed,
                 ).isEqualTo(true)
             }
         }
